rename files (part 1) to remove .ethers
