+++ unlock-js/src/__tests__/helpers/walletServiceHelper.js
