+++ unlock-js/src/__tests__/v0/partialWithdrawFromLock.test.js
@@ -1,14 +1,11 @@
-import * as UnlockV02 from 'unlock-abi-0-2'
+import * as UnlockV0 from 'unlock-abi-0'
 import * as utils from '../../utils.ethers'
 import Errors from '../../errors'
 import TransactionTypes from '../../transactionTypes'
 import NockHelper from '../helpers/nockHelper'
-import {
-  prepWalletService,
-  prepContract,
-} from '../helpers/walletServiceHelper.ethers'
+import { prepWalletService, prepContract } from '../helpers/walletServiceHelper'
 
-const { FAILED_TO_UPDATE_KEY_PRICE } = Errors
+const { FAILED_TO_WITHDRAW_FROM_LOCK } = Errors
 const endpoint = 'http://127.0.0.1:8545'
 const nock = new NockHelper(endpoint, false /** debug */, true /** ethers */)
 
@@ -18,23 +15,23 @@ let transactionResult
 let setupSuccess
 let setupFail
 
-describe('v02', () => {
-  describe('updateKeyPrice', () => {
-    const lockAddress = '0xd8c88be5e8eb88e38e6ff5ce186d764676012b0b'
+describe('v0', () => {
+  describe('partialWithdrawFromLock', () => {
+    const lock = '0xd8c88be5e8eb88e38e6ff5ce186d764676012b0b'
     const account = '0xdeadbeef'
-    const price = '100000000'
+    const amount = '3'
 
     async function nockBeforeEach() {
       nock.cleanAll()
       walletService = await prepWalletService(
-        UnlockV02.PublicLock,
+        UnlockV0.PublicLock,
         endpoint,
         nock
       )
 
       const callMethodData = prepContract({
-        contract: UnlockV02.PublicLock,
-        functionName: 'updateKeyPrice',
+        contract: UnlockV0.PublicLock,
+        functionName: 'partialWithdraw',
         signature: 'uint256',
         nock,
       })
@@ -44,7 +41,7 @@ describe('v02', () => {
         testTransactionResult,
         success,
         fail,
-      } = callMethodData(utils.toWei(price, 'ether'))
+      } = callMethodData(utils.toWei(amount, 'ether'))
 
       transaction = testTransaction
       transactionResult = testTransactionResult
@@ -54,6 +51,7 @@ describe('v02', () => {
 
     it('should invoke _handleMethodCall with the right params', async () => {
       expect.assertions(2)
+      const callback = jest.fn()
 
       await nockBeforeEach()
       setupSuccess()
@@ -63,11 +61,16 @@ describe('v02', () => {
       )
       const mock = walletService._handleMethodCall
 
-      await walletService.updateKeyPrice(lockAddress, account, price)
+      await walletService.partialWithdrawFromLock(
+        lock,
+        account,
+        amount,
+        callback
+      )
 
       expect(mock).toHaveBeenCalledWith(
         expect.any(Promise),
-        TransactionTypes.UPDATE_KEY_PRICE
+        TransactionTypes.WITHDRAWAL
       )
 
       // verify that the promise passed to _handleMethodCall actually resolves
@@ -77,19 +80,47 @@ describe('v02', () => {
       await nock.resolveWhenAllNocksUsed()
     })
 
-    it('should emit an error if the transaction could not be sent', async () => {
+    it('should emit an error if the transaction cannot be sent', async () => {
       expect.assertions(1)
+      const callback = jest.fn()
 
       const error = { code: 404, data: 'oops' }
       await nockBeforeEach()
       setupFail(error)
 
       walletService.on('error', error => {
-        expect(error.message).toBe(FAILED_TO_UPDATE_KEY_PRICE)
+        expect(error.message).toBe(FAILED_TO_WITHDRAW_FROM_LOCK)
       })
 
-      await walletService.updateKeyPrice(lockAddress, account, price)
+      await walletService.partialWithdrawFromLock(
+        lock,
+        account,
+        amount,
+        callback
+      )
       await nock.resolveWhenAllNocksUsed()
+    })
+
+    it('should not emit an error when `error` is falsy', async () => {
+      expect.assertions(1)
+      const callback = jest.fn()
+
+      await nockBeforeEach()
+      setupSuccess()
+
+      walletService._handleMethodCall = jest.fn(() =>
+        Promise.resolve(transaction.hash)
+      )
+
+      await walletService.partialWithdrawFromLock(
+        lock,
+        account,
+        amount,
+        callback
+      )
+
+      await nock.resolveWhenAllNocksUsed()
+      expect(callback).toHaveBeenCalled()
     })
   })
 })
