+++ unlock-js/src/__tests__/v02/createLock.test.js
@@ -1,14 +1,12 @@
-import * as UnlockV0 from 'unlock-abi-0'
+import { ethers } from 'ethers'
+import * as UnlockV02 from 'unlock-abi-0-2'
 import * as utils from '../../utils.ethers'
 import Errors from '../../errors'
 import TransactionTypes from '../../transactionTypes'
 import NockHelper from '../helpers/nockHelper'
-import {
-  prepWalletService,
-  prepContract,
-} from '../helpers/walletServiceHelper.ethers'
+import { prepWalletService, prepContract } from '../helpers/walletServiceHelper'
 
-const { FAILED_TO_PURCHASE_KEY } = Errors
+const { FAILED_TO_CREATE_LOCK } = Errors
 const endpoint = 'http://127.0.0.1:8545'
 const nock = new NockHelper(endpoint, false /** debug */, true /** ethers */)
 
@@ -17,29 +15,30 @@ let transaction
 let transactionResult
 let setupSuccess
 let setupFail
+const lock = {
+  address: '0x0987654321098765432109876543210987654321',
+  expirationDuration: 86400, // 1 day
+  keyPrice: '0.1', // 0.1 Eth
+  maxNumberOfKeys: 100,
+}
+const owner = '0xdeadfeed'
 
-describe('v0', () => {
-  describe('purchaseKey', () => {
-    const keyPrice = '0.01'
-    const owner = '0xab7c74abc0c4d48d1bdad5dcb26153fc8780f83e'
-    const lockAddress = '0xd8c88be5e8eb88e38e6ff5ce186d764676012b0b'
-    const account = '0xdeadbeef'
-    const data = 'key data'
-
+describe('v02', () => {
+  describe('createLock', () => {
     async function nockBeforeEach() {
       nock.cleanAll()
       walletService = await prepWalletService(
-        UnlockV0.PublicLock,
+        UnlockV02.Unlock,
         endpoint,
-        nock
+        nock,
+        true // this is the Unlock contract, not PublicLock
       )
 
       const callMethodData = prepContract({
-        contract: UnlockV0.PublicLock,
-        functionName: 'purchaseFor',
-        signature: 'address,bytes',
+        contract: UnlockV02.Unlock,
+        functionName: 'createLock',
+        signature: 'uint256,address,uint256,uint256',
         nock,
-        value: keyPrice,
       })
 
       const {
@@ -47,7 +46,12 @@ describe('v0', () => {
         testTransactionResult,
         success,
         fail,
-      } = callMethodData(owner, utils.utf8ToHex(data))
+      } = callMethodData(
+        lock.expirationDuration,
+        ethers.constants.AddressZero,
+        utils.toWei(lock.keyPrice, 'ether'),
+        lock.maxNumberOfKeys
+      )
 
       transaction = testTransaction
       transactionResult = testTransactionResult
@@ -66,23 +70,40 @@ describe('v0', () => {
       )
       const mock = walletService._handleMethodCall
 
-      await walletService.purchaseKey(
-        lockAddress,
-        owner,
-        keyPrice,
-        account,
-        data
-      )
+      await walletService.createLock(lock, owner)
 
       expect(mock).toHaveBeenCalledWith(
         expect.any(Promise),
-        TransactionTypes.KEY_PURCHASE
+        TransactionTypes.LOCK_CREATION
       )
 
       // verify that the promise passed to _handleMethodCall actually resolves
       // to the result the chain returns from a sendTransaction call to createLock
       const result = await mock.mock.calls[0][0]
       expect(result).toEqual(transactionResult)
+      await nock.resolveWhenAllNocksUsed()
+    })
+
+    it('should emit lock.updated with the transaction', async () => {
+      expect.assertions(2)
+
+      await nockBeforeEach()
+      setupSuccess()
+
+      walletService.on('lock.updated', (lockAddress, update) => {
+        expect(lockAddress).toBe(lock.address)
+        expect(update).toEqual({
+          transaction: transaction.hash,
+          balance: '0',
+          expirationDuration: lock.expirationDuration,
+          keyPrice: lock.keyPrice,
+          maxNumberOfKeys: lock.maxNumberOfKeys,
+          outstandingKeys: 0,
+          owner,
+        })
+      })
+
+      await walletService.createLock(lock, owner)
       await nock.resolveWhenAllNocksUsed()
     })
 
@@ -94,16 +115,10 @@ describe('v0', () => {
       setupFail(error)
 
       walletService.on('error', error => {
-        expect(error.message).toBe(FAILED_TO_PURCHASE_KEY)
+        expect(error.message).toBe(FAILED_TO_CREATE_LOCK)
       })
 
-      await walletService.purchaseKey(
-        lockAddress,
-        owner,
-        keyPrice,
-        account,
-        data
-      )
+      await walletService.createLock(lock, owner)
       await nock.resolveWhenAllNocksUsed()
     })
   })
