+++ unlock-js/src/__tests__/v02/withdrawFromLock.test.js
@@ -1,15 +1,10 @@
-import { ethers } from 'ethers'
-import * as UnlockV01 from 'unlock-abi-0-1'
-import * as utils from '../../utils.ethers'
+import * as UnlockV02 from 'unlock-abi-0-2'
 import Errors from '../../errors'
 import TransactionTypes from '../../transactionTypes'
 import NockHelper from '../helpers/nockHelper'
-import {
-  prepWalletService,
-  prepContract,
-} from '../helpers/walletServiceHelper.ethers'
+import { prepWalletService, prepContract } from '../helpers/walletServiceHelper'
 
-const { FAILED_TO_CREATE_LOCK } = Errors
+const { FAILED_TO_WITHDRAW_FROM_LOCK } = Errors
 const endpoint = 'http://127.0.0.1:8545'
 const nock = new NockHelper(endpoint, false /** debug */, true /** ethers */)
 
@@ -18,29 +13,24 @@ let transaction
 let transactionResult
 let setupSuccess
 let setupFail
-const lock = {
-  address: '0x0987654321098765432109876543210987654321',
-  expirationDuration: 86400, // 1 day
-  keyPrice: '0.1', // 0.1 Eth
-  maxNumberOfKeys: 100,
-}
-const owner = '0xdeadfeed'
 
-describe('v01', () => {
-  describe('createLock', () => {
+describe('v02', () => {
+  describe('withdrawFromLock', () => {
+    const lockAddress = '0xd8c88be5e8eb88e38e6ff5ce186d764676012b0b'
+    const account = '0xdeadbeef'
+
     async function nockBeforeEach() {
       nock.cleanAll()
       walletService = await prepWalletService(
-        UnlockV01.Unlock,
+        UnlockV02.PublicLock,
         endpoint,
-        nock,
-        true // this is the Unlock contract, not PublicLock
+        nock
       )
 
       const callMethodData = prepContract({
-        contract: UnlockV01.Unlock,
-        functionName: 'createLock',
-        signature: 'uint256,address,uint256,uint256',
+        contract: UnlockV02.PublicLock,
+        functionName: 'withdraw',
+        signature: '',
         nock,
       })
 
@@ -49,12 +39,7 @@ describe('v01', () => {
         testTransactionResult,
         success,
         fail,
-      } = callMethodData(
-        lock.expirationDuration,
-        ethers.constants.AddressZero,
-        utils.toWei(lock.keyPrice, 'ether'),
-        lock.maxNumberOfKeys
-      )
+      } = callMethodData()
 
       transaction = testTransaction
       transactionResult = testTransactionResult
@@ -73,40 +58,17 @@ describe('v01', () => {
       )
       const mock = walletService._handleMethodCall
 
-      await walletService.createLock(lock, owner)
+      await walletService.withdrawFromLock(lockAddress, account)
 
       expect(mock).toHaveBeenCalledWith(
         expect.any(Promise),
-        TransactionTypes.LOCK_CREATION
+        TransactionTypes.WITHDRAWAL
       )
 
       // verify that the promise passed to _handleMethodCall actually resolves
       // to the result the chain returns from a sendTransaction call to createLock
       const result = await mock.mock.calls[0][0]
       expect(result).toEqual(transactionResult)
-      await nock.resolveWhenAllNocksUsed()
-    })
-
-    it('should emit lock.updated with the transaction', async () => {
-      expect.assertions(2)
-
-      await nockBeforeEach()
-      setupSuccess()
-
-      walletService.on('lock.updated', (lockAddress, update) => {
-        expect(lockAddress).toBe(lock.address)
-        expect(update).toEqual({
-          transaction: transaction.hash,
-          balance: '0',
-          expirationDuration: lock.expirationDuration,
-          keyPrice: lock.keyPrice,
-          maxNumberOfKeys: lock.maxNumberOfKeys,
-          outstandingKeys: 0,
-          owner,
-        })
-      })
-
-      await walletService.createLock(lock, owner)
       await nock.resolveWhenAllNocksUsed()
     })
 
@@ -118,10 +80,9 @@ describe('v01', () => {
       setupFail(error)
 
       walletService.on('error', error => {
-        expect(error.message).toBe(FAILED_TO_CREATE_LOCK)
+        expect(error.message).toBe(FAILED_TO_WITHDRAW_FROM_LOCK)
       })
-
-      await walletService.createLock(lock, owner)
+      await walletService.withdrawFromLock(lockAddress, account)
       await nock.resolveWhenAllNocksUsed()
     })
   })
