+++ unlock-js/src/__tests__/walletService.test.js
