+++ unlock-js/src/__tests__/web3Service.test.js
